requirements.txt
pyautogen>=0.2.3
fastapi
psycopg2-binary
uvicorn[standard]
